Pull request #590: Migrate to Conan 2.0
Merge in ADGUARD-CORE-LIBS/dns-libs from feature/migrate_to_conan_2.0_syntax to master

Squashed commit of the following:

commit a32a50f5682b7a0e05b8212108f30dd3b5afcd0b
Author: Max Grupper <m.grupper@adguard.com>
Date:   Fri Dec 1 11:15:47 2023 +0200

    skipci: redundant asterix

commit 3ff13923ff7b9fdd8ad8088bb5effee4b2bde3e5
Author: Max Grupper <m.grupper@adguard.com>
Date:   Fri Dec 1 11:06:01 2023 +0200

    skipci: bump nlc version

commit bb080e276de10e5ca6c585fb7d4e2c25418d68c4
Author: Max Grupper <m.grupper@adguard.com>
Date:   Fri Dec 1 10:49:41 2023 +0200

    skipci: missing magic_enum & fix version

commit 283d9a22cb8213a0e4aed1c6e02a0b22ee86f4ef
Merge: 3cf3f5f2 50b25d54
Author: Max Grupper <m.grupper@adguard.com>
Date:   Fri Dec 1 10:44:47 2023 +0200

    Merge branch 'master' into feature/migrate_to_conan_2.0_syntax

commit 3cf3f5f26ad1378cff63d92a565436d06f998c50
Author: Max Grupper <m.grupper@adguard.com>
Date:   Fri Dec 1 10:41:11 2023 +0200

    remove temp

commit 660aa3c9213dc42aa7fc3ac681fad19c3ecd024c
Author: Max Grupper <m.grupper@adguard.com>
Date:   Thu Nov 30 10:42:34 2023 +0200

    skipci: magic enum

commit 501855c37dbea3e25f6e4716700628be3b2606b1
Author: Max Grupper <m.grupper@adguard.com>
Date:   Thu Nov 30 10:29:37 2023 +0200

    skipci: fix python syntax

commit 7478ead0f12f9e45b5be19b09196dc652820d79f
Author: Sergey Fionov <sfionov@adguard.com>
Date:   Thu Nov 30 10:20:54 2023 +0200

    Fix conan upload

commit da854ccb3b21ce59bbe10849f8e230ed88615abb
Author: Max Grupper <m.grupper@adguard.com>
Date:   Wed Nov 29 22:32:42 2023 +0200

    skipci: append detours for windows to requires

commit 4bbab63cdeb2e2f346cc6dee90a4d8471ea28b44
Author: Max Grupper <m.grupper@adguard.com>
Date:   Wed Nov 29 17:09:29 2023 +0200

    skipci: use ssh in vcs_url

commit 545811c1a4c23fc70b54d6bf454571f53f72fa73
Author: Max Grupper <m.grupper@adguard.com>
Date:   Wed Nov 29 11:49:10 2023 +0200

    skipci: ready for merge

commit 5301a06deac140134823fe77d9c479c5fc3e83bb
Author: Max Grupper <m.grupper@adguard.com>
Date:   Wed Nov 29 11:47:22 2023 +0200

    skipci: useless conan install

commit 969257ff073aecabbd3af62685f8dac39b4b2cd6
Author: Max Grupper <m.grupper@adguard.com>
Date:   Wed Nov 29 11:39:29 2023 +0200

    skipci: undo adding local files

commit 5d5a6da7e8bb87d0b4d6a7de48225b2587e41b0b
Merge: 7fa480c9 33cd4453
Author: Max Grupper <m.grupper@adguard.com>
Date:   Wed Nov 29 11:34:01 2023 +0200

    Merge remote-tracking branch 'origin/feature/migrate_to_conan_2.0_syntax' into feature/migrate_to_conan_2.0_syntax

commit 7fa480c96513f104d15b550f961afefa01492eb8
Author: Max Grupper <m.grupper@adguard.com>
Date:   Wed Nov 29 11:33:28 2023 +0200

    win: create venv in root

commit 33cd4453de6a15f06cca47a343dc0f7bee85eac1
Author: Max Grupper <m.grupper@adguard.com>
Date:   Wed Nov 29 11:05:48 2023 +0200

    add conan provider patch

commit 44d30cf1e74e5e57737a9c7fe3a3d6c174651545
Author: Max Grupper <m.grupper@adguard.com>
Date:   Wed Nov 29 10:50:58 2023 +0200

    win: fix

commit a36ea8cea848dabea31dbc860ca631343f1bb552
Author: Sergey Fionov <sfionov@adguard.com>
Date:   Tue Nov 28 23:09:27 2023 +0200

    Downgrade boringssl

commit ae13df48fb9e844dc7ff0b4965abff47fd8b0d49
Author: Sergey Fionov <sfionov@adguard.com>
Date:   Tue Nov 28 22:18:50 2023 +0200

    Fix visibility collisions

commit 4929852bb9cba55ada5b1f146fc466e3ec6d763a
Author: Sergey Fionov <sfionov@adguard.com>
Date:   Tue Nov 28 21:40:27 2023 +0200

    Fix build

... and 35 more commits

+++ proxy/response_cache.h
@@ -191,8 +191,8 @@ private:
     static std::string get_cache_key(const ldns_pkt *request) {
         const auto *question = ldns_rr_list_rr(ldns_pkt_question(request), 0);
         std::string key = fmt::format("{}|{}|{}{}|", // '|' is to avoid collisions
-                ldns_rr_get_type(question), ldns_rr_get_class(question), ldns_pkt_edns_do(request) ? "1" : "0",
-                ldns_pkt_cd(request) ? "1" : "0");
+                (int) ldns_rr_get_type(question), (int) ldns_rr_get_class(question),
+                ldns_pkt_edns_do(request) ? "1" : "0", ldns_pkt_cd(request) ? "1" : "0");
 
         // Compute the domain name, in lower case for case-insensitivity
         const auto *owner = ldns_rr_owner(question);
